Add (git) to package description
In the table of package managers, the AUR packages for both the stable + `-git` versons were listed as `Arch Linux`.
Though, `Arch Linux (Git)` is not a distro, it seems to read as "version for". Their locations have also been swapped, because `master` can have bugs sometimes, and most users will find the first and not the second.
And so this fixes it, indicating that the second is a git version.